+++ CHANGELOG.md
@@ -4,5 +4,8 @@
 
 * Read vector length upfront, rather than on every access.
   * `FlatBuffers.Vector.length` changed from `Either ReadError Int32` to `Int32`.
+* Added:
+  * `FlatBuffers.Vector.take`
+  * `FlatBuffers.Vector.drop`